Add Danmaku-related Toggle and Inputs
Add two optional inputs for DanmakuFactory command used in session.py and a toggle for burning video with high-energy bar and danmaku in record-only mode. All three arguments can be configured in recorder_config.yaml under 'rooms' seciton, an example config file is added.

Backward Compatibility: Default values are added to recorder_config.py so that nothing will be affected if above three arguemnts are not specified in config file.